+++ mvvmfx/src/main/java/de/saxsys/mvvmfx/utils/validation/CompositeValidator.java
@@ -1,7 +1,4 @@
-package de.saxsys.mvvmfx.utils.validation.validators;
-
-import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
-import de.saxsys.mvvmfx.utils.validation.Validator;
+package de.saxsys.mvvmfx.utils.validation;
 
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
