+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/utils/validation/RulesTest.java
@@ -1,6 +1,5 @@
-package de.saxsys.mvvmfx.utils.validation.validators;
+package de.saxsys.mvvmfx.utils.validation;
 
-import de.saxsys.mvvmfx.utils.validation.validators.ObservableRules;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.beans.value.ObservableBooleanValue;
@@ -20,7 +19,7 @@ public class RulesTest {
 	public void testNotNullOrEmpty() {
 		StringProperty value = new SimpleStringProperty();
 
-		final ObservableBooleanValue rule = ObservableRules.notEmpty(value);
+		final ObservableBooleanValue rule = Rules.notEmpty(value);
 		
 		assertThat(rule.get()).isFalse();
 
@@ -48,7 +47,7 @@ public class RulesTest {
 	public void testMatches() {
 		StringProperty value = new SimpleStringProperty();
 
-		final ObservableBooleanValue rule = ObservableRules.matches(value, Pattern.compile("[0-9]"));
+		final ObservableBooleanValue rule = Rules.matches(value, Pattern.compile("[0-9]"));
 		
 		assertThat(rule.get()).isFalse();
